+++ packages/plugin-faker/src/generators/__snapshots__/petWithDateString.ts
@@ -2,7 +2,7 @@ import { faker } from "@faker-js/faker";
 
  export function pet(data?: Partial<Pet>) {
     return {
-        ...{ "id": faker.number.int(), "name": faker.string.alpha(), "tag": faker.string.alpha(), "code": faker.helpers.arrayElement<any>([faker.string.alpha(), faker.helpers.fromRegExp(new RegExp("\\b[1-9]\\b"))]), "shipDate": faker.date.anytime().toString(), "shipTime": faker.date.anytime().toString() },
+        ...{ "id": faker.number.int(), "name": faker.string.alpha(), "tag": faker.string.alpha(), "code": faker.helpers.fromRegExp(new RegExp("\\b[1-9]\\b")), "shipDate": faker.date.anytime().toString(), "shipTime": faker.date.anytime().toString() },
         ...data || {}
     };
 }