+++ docs/changelog.md
@@ -5,6 +5,7 @@ title: Changelog
 # Changelog
 
 ## 3.0.1
+- [`plugin-faker`](/plugins/plugin-faker): Correct faker functions for uuid, pattern and email
 - [`plugin-react-query`](/plugins/plugin-react-query/): allow disabling `useQuery`
 - [`plugin-react-query`](/plugins/plugin-react-query/): use of `InfiniteData` TypeScript helper for infiniteQueries
 - [`plugin-vue-query`](/plugins/plugin-vue-query/): use of `InfiniteData` TypeScript helper for infiniteQueries